Refactored Disambiguation
